+++ lib/app/modules/single_shop/transaction_and_refund/data/remote/models/quick_sell_request.dart
@@ -4,9 +4,11 @@
 
 import 'dart:convert';
 
-QuickSellRequest quickSellRequestFromJson(String str) => QuickSellRequest.fromJson(json.decode(str));
+QuickSellRequest quickSellRequestFromJson(String str) =>
+    QuickSellRequest.fromJson(json.decode(str));
 
-String quickSellRequestToJson(QuickSellRequest data) => json.encode(data.toJson());
+String quickSellRequestToJson(QuickSellRequest data) =>
+    json.encode(data.toJson());
 
 class QuickSellRequest {
   QuickSellRequest({
@@ -25,7 +27,7 @@ class QuickSellRequest {
   int shopId;
   String createdAt;
   String updatedAt;
-  int price;
+  double price;
   String details;
   String image;
   String uniqueId;
@@ -55,32 +57,36 @@ class QuickSellRequest {
         uniqueId: uniqueId ?? this.uniqueId,
         version: version ?? this.version,
         transactionUniqueId: transactionUniqueId ?? this.transactionUniqueId,
-        profit: profit??this.profit,
+        profit: profit ?? this.profit,
       );
 
-  factory QuickSellRequest.fromJson(Map<String, dynamic> json) => QuickSellRequest(
-    shopId: json["shop_id"] == null ? null : json["shop_id"],
-    createdAt: json["created_at"] == null ? null : json["created_at"],
-    updatedAt: json["updated_at"] == null ? null : json["updated_at"],
-    price: json["price"] == null ? null : json["price"],
-    details: json["details"] == null ? null : json["details"],
-    image: json["image"] == null ? null : json["image"],
-    uniqueId: json["unique_id"] == null ? null : json["unique_id"],
-    version: json["version"] == null ? null : json["version"],
-    profit: json["profit"] == null ? null : json["profit"],
-    transactionUniqueId: json["transaction_unique_id"] == null ? null : json["transaction_unique_id"],
-  );
+  factory QuickSellRequest.fromJson(Map<String, dynamic> json) =>
+      QuickSellRequest(
+        shopId: json["shop_id"] == null ? null : json["shop_id"],
+        createdAt: json["created_at"] == null ? null : json["created_at"],
+        updatedAt: json["updated_at"] == null ? null : json["updated_at"],
+        price: json["price"] == null ? null : json["price"],
+        details: json["details"] == null ? null : json["details"],
+        image: json["image"] == null ? null : json["image"],
+        uniqueId: json["unique_id"] == null ? null : json["unique_id"],
+        version: json["version"] == null ? null : json["version"],
+        profit: json["profit"] == null ? null : json["profit"],
+        transactionUniqueId: json["transaction_unique_id"] == null
+            ? null
+            : json["transaction_unique_id"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "shop_id": shopId == null ? null : shopId,
-    "created_at": createdAt == null ? null : createdAt,
-    "updated_at": updatedAt == null ? null : updatedAt,
-    "price": price == null ? null : price,
-    "details": details == null ? null : details,
-    "image": image == null ? null : image,
-    "unique_id": uniqueId == null ? null : uniqueId,
-    "version": version == null ? null : version,
-    "transaction_unique_id": transactionUniqueId == null ? null : transactionUniqueId,
-    "profit": profit == null ? null : profit,
-  };
+        "shop_id": shopId == null ? null : shopId,
+        "created_at": createdAt == null ? null : createdAt,
+        "updated_at": updatedAt == null ? null : updatedAt,
+        "price": price == null ? null : price,
+        "details": details == null ? null : details,
+        "image": image == null ? null : image,
+        "unique_id": uniqueId == null ? null : uniqueId,
+        "version": version == null ? null : version,
+        "transaction_unique_id":
+            transactionUniqueId == null ? null : transactionUniqueId,
+        "profit": profit == null ? null : profit,
+      };
 }
